requirements.txt
# ---- Core Framework ----
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic[email]>=2.5.0
email-validator>=2.0.0
pydantic-settings>=2.1.0

# ---- Database ----
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0

# ---- Auth ----
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
python-multipart>=0.0.6

# ---- Task Queue ----
celery[rabbitmq]>=5.3.0
redis>=5.0.0

# ---- LLM Providers ----
openai>=1.12.0
google-generativeai>=0.4.0
httpx>=0.27.0

# ---- Configuration ----
pyyaml>=6.0.1
python-dotenv>=1.0.0

# ---- Logging ----
python-json-logger>=2.0.7

# ---- File handling ----
aiofiles>=23.2.0
